+++ apps/examples/dpp/main.c
@@ -122,7 +122,7 @@ static void ticker_task(void *param) {
 int main(void) {
     am_ao_state_ctor(/*cfg=*/NULL);
 
-    am_event_add_pool(
+    am_event_pool_add(
         m_event_pool,
         sizeof(m_event_pool),
         sizeof(m_event_pool[0]),