+++ apps/examples/smokers/main.c
@@ -338,7 +338,7 @@ AM_ALIGNOF_DEFINE(events_t);
 int main(void) {
     am_ao_state_ctor(/*cfg=*/NULL);
 
-    am_event_add_pool(
+    am_event_pool_add(
         m_event_pool,
         sizeof(m_event_pool),
         sizeof(m_event_pool[0]),