+++ libs/ao/tests/publish.c
@@ -95,7 +95,7 @@ static void test_publish(void) {
     };
     am_ao_state_ctor(&cfg);
 
-    am_event_add_pool(
+    am_event_pool_add(
         m_event_pool,
         sizeof(m_event_pool),
         sizeof(m_event_pool[0]),