+++ libs/hsm/tests/event_queue.c
@@ -159,7 +159,7 @@ int main(void) {
         static char pool[1 * AM_POOL_BLOCK_SIZEOF(struct am_event)] AM_ALIGNED(
             AM_ALIGN_MAX
         );
-        am_event_add_pool(
+        am_event_pool_add(
             pool,
             (int)sizeof(pool),
             AM_POOL_BLOCK_SIZEOF(struct am_event),