+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [0-based versioning](https://0ver.org/).
 - Make `am_event_queue_get_nfree_min()` thread safe.
 - Remove `am_event_queue_get_nfree()`.
 - Rename `am_event_queue_get_nbusy()` to `am_event_queue_get_nbusy_unsafe()`.
+- Rename `am_event_add_pool()` to  `am_event_pool_add()`.
 
 ### Added